refactor: 예약 생성 관련 코드 하나로 통합
