--- src/main/java/roomescape/service/reservation/ReservationCreateValidator.java
@@ -1,64 +1,0 @@
-package roomescape.service.reservation;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import org.springframework.stereotype.Component;
-import roomescape.domain.Member;
-import roomescape.domain.ReservationTime;
-import roomescape.domain.Theme;
-import roomescape.repository.MemberRepository;
-import roomescape.repository.ReservationRepository;
-import roomescape.repository.ReservationTimeRepository;
-import roomescape.repository.ThemeRepository;
-
-@Component
-public class ReservationCreateValidator {
-
-    private final ReservationRepository reservationRepository;
-    private final ReservationTimeRepository reservationTimeRepository;
-    private final ThemeRepository themeRepository;
-    private final MemberRepository memberRepository;
-
-    public ReservationCreateValidator(ReservationRepository reservationRepository,
-                                      ReservationTimeRepository reservationTimeRepository,
-                                      ThemeRepository themeRepository,
-                                      MemberRepository memberRepository) {
-        this.reservationRepository = reservationRepository;
-        this.reservationTimeRepository = reservationTimeRepository;
-        this.themeRepository = themeRepository;
-        this.memberRepository = memberRepository;
-    }
-
-
-    public ReservationTime getValidReservationTime(long reservationId) {
-        return reservationTimeRepository.findById(reservationId)
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 예약 시간 입니다."));
-    }
-
-    public Theme getValidTheme(long themeId) {
-        return themeRepository.findById(themeId)
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 테마 입니다."));
-    }
-
-    public void validateAlreadyBooked(LocalDate date, long timeId, long themeId) {
-        if (reservationRepository.existsByDateAndTimeIdAndThemeId(date, timeId, themeId)) {
-            throw new IllegalArgumentException("해당 시간에 이미 예약된 테마입니다.");
-        }
-    }
-
-    public void validateDateIsFuture(LocalDate date, ReservationTime reservationTime) {
-        LocalDateTime localDateTime = toLocalDateTime(date, reservationTime);
-        if (localDateTime.isBefore(LocalDateTime.now())) {
-            throw new IllegalArgumentException("지나간 날짜와 시간에 대한 예약 생성은 불가능합니다.");
-        }
-    }
-
-    private LocalDateTime toLocalDateTime(LocalDate date, ReservationTime reservationTime) {
-        return LocalDateTime.of(date, reservationTime.getStartAt());
-    }
-
-    public Member getValidMember(long memberId) {
-        return memberRepository.findById(memberId)
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 회원입니다."));
-    }
-}
